Split encode/encode_value
