+++ common/searchattribute/encode_test.go
@@ -30,103 +30,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	enumspb "go.temporal.io/api/enums/v1"
-	"go.temporal.io/sdk/converter"
-
-	"go.temporal.io/server/common/payload"
 )
-
-func Test_DecodeValue_FromMetadata_Success(t *testing.T) {
-	assert := assert.New(t)
-
-	payloadStr := payload.EncodeString("qwe")
-	payloadStr.Metadata["type"] = []byte("String")
-	decodedStr, err := DecodeValue(payloadStr, enumspb.INDEXED_VALUE_TYPE_UNSPECIFIED)
-	assert.NoError(err)
-	assert.Equal("qwe", decodedStr)
-
-	payloadInt, err := payload.Encode(123)
-	assert.NoError(err)
-	payloadInt.Metadata["type"] = []byte("Int")
-	decodedInt, err := DecodeValue(payloadInt, enumspb.INDEXED_VALUE_TYPE_STRING) // MetadataType should be used anyway
-	assert.NoError(err)
-	assert.Equal(int64(123), decodedInt)
-
-	payloadBool, err := payload.Encode(true)
-	assert.NoError(err)
-	payloadBool.Metadata["type"] = []byte("Bool")
-	decodedBool, err := DecodeValue(payloadBool, enumspb.INDEXED_VALUE_TYPE_UNSPECIFIED)
-	assert.NoError(err)
-	assert.Equal(true, decodedBool)
-}
-
-func Test_DecodeValue_FromParameter_Success(t *testing.T) {
-	assert := assert.New(t)
-
-	payloadStr := payload.EncodeString("qwe")
-	decodedStr, err := DecodeValue(payloadStr, enumspb.INDEXED_VALUE_TYPE_STRING)
-	assert.NoError(err)
-	assert.Equal("qwe", decodedStr)
-
-	payloadInt, err := payload.Encode(123)
-	assert.NoError(err)
-	decodedInt, err := DecodeValue(payloadInt, enumspb.INDEXED_VALUE_TYPE_INT)
-	assert.NoError(err)
-	assert.Equal(int64(123), decodedInt)
-
-	payloadInt, err = payload.Encode(123)
-	assert.NoError(err)
-	payloadInt.Metadata["type"] = []byte("UnknownType") // should not be used because incorrect
-	decodedInt, err = DecodeValue(payloadInt, enumspb.INDEXED_VALUE_TYPE_INT)
-	assert.NoError(err)
-	assert.Equal(int64(123), decodedInt)
-
-	payloadBool, err := payload.Encode(true)
-	assert.NoError(err)
-	decodedBool, err := DecodeValue(payloadBool, enumspb.INDEXED_VALUE_TYPE_BOOL)
-	assert.NoError(err)
-	assert.Equal(true, decodedBool)
-}
-
-func Test_DecodeValue_Error(t *testing.T) {
-	assert := assert.New(t)
-
-	payloadStr := payload.EncodeString("qwe")
-	decodedStr, err := DecodeValue(payloadStr, enumspb.INDEXED_VALUE_TYPE_UNSPECIFIED)
-	assert.Error(err)
-	assert.True(errors.Is(err, ErrInvalidType))
-	assert.Nil(decodedStr)
-
-	payloadInt, err := payload.Encode(123)
-	assert.NoError(err)
-	payloadInt.Metadata["type"] = []byte("UnknownType")
-	decodedInt, err := DecodeValue(payloadInt, enumspb.INDEXED_VALUE_TYPE_UNSPECIFIED)
-	assert.Error(err)
-	assert.True(errors.Is(err, ErrInvalidType))
-	assert.Nil(decodedInt)
-
-	payloadInt, err = payload.Encode(123)
-	assert.NoError(err)
-	payloadInt.Metadata["type"] = []byte("String")
-	decodedInt, err = DecodeValue(payloadInt, enumspb.INDEXED_VALUE_TYPE_UNSPECIFIED)
-	assert.Error(err)
-	assert.True(errors.Is(err, converter.ErrUnableToDecode), err.Error())
-	assert.Nil(decodedInt)
-}
-
-func Test_EncodeValue(t *testing.T) {
-	assert := assert.New(t)
-
-	encodedPayload, err := EncodeValue(123, enumspb.INDEXED_VALUE_TYPE_INT)
-	assert.NoError(err)
-	assert.Equal("123", string(encodedPayload.GetData()))
-	assert.Equal("Int", string(encodedPayload.Metadata["type"]))
-
-	encodedPayload, err = EncodeValue("qwe", enumspb.INDEXED_VALUE_TYPE_STRING)
-	assert.NoError(err)
-	assert.Equal(`"qwe"`, string(encodedPayload.GetData()))
-	assert.Equal("String", string(encodedPayload.Metadata["type"]))
-
-}
 
 func Test_Encode_Success(t *testing.T) {
 	assert := assert.New(t)
@@ -185,4 +89,86 @@ func Test_Encode_Error(t *testing.T) {
 	assert.Equal("2", string(sa.IndexedFields["key2"].GetData()))
 	assert.Equal("true", string(sa.IndexedFields["key3"].GetData()))
 	assert.Equal("Bool", string(sa.IndexedFields["key3"].GetMetadata()["type"]))
+}
+
+func Test_Decode_Success(t *testing.T) {
+	assert := assert.New(t)
+
+	typeMap := map[string]enumspb.IndexedValueType{
+		"key1": enumspb.INDEXED_VALUE_TYPE_STRING,
+		"key2": enumspb.INDEXED_VALUE_TYPE_INT,
+		"key3": enumspb.INDEXED_VALUE_TYPE_BOOL,
+	}
+	sa, err := Encode(map[string]interface{}{
+		"key1": "val1",
+		"key2": 2,
+		"key3": true,
+	}, typeMap)
+	assert.NoError(err)
+
+	vals, err := Decode(sa, typeMap)
+	assert.NoError(err)
+	assert.Len(vals, 3)
+	assert.Equal("val1", vals["key1"])
+	assert.Equal(int64(2), vals["key2"])
+	assert.Equal(true, vals["key3"])
+
+	delete(sa.IndexedFields["key1"].Metadata, "type")
+	delete(sa.IndexedFields["key2"].Metadata, "type")
+	delete(sa.IndexedFields["key3"].Metadata, "type")
+
+	vals, err = Decode(sa, typeMap)
+	assert.NoError(err)
+	assert.Len(vals, 3)
+	assert.Equal("val1", vals["key1"])
+	assert.Equal(int64(2), vals["key2"])
+	assert.Equal(true, vals["key3"])
+}
+
+func Test_Decode_Error(t *testing.T) {
+	assert := assert.New(t)
+
+	typeMap := map[string]enumspb.IndexedValueType{
+		"key1": enumspb.INDEXED_VALUE_TYPE_STRING,
+		"key2": enumspb.INDEXED_VALUE_TYPE_INT,
+		"key3": enumspb.INDEXED_VALUE_TYPE_BOOL,
+	}
+	sa, err := Encode(map[string]interface{}{
+		"key1": "val1",
+		"key2": 2,
+		"key3": true,
+	}, typeMap)
+	assert.NoError(err)
+
+	vals, err := Decode(sa, nil)
+	assert.Error(err)
+	assert.True(errors.Is(err, ErrTypeMapIsEmpty))
+	assert.Len(sa.IndexedFields, 3)
+	assert.Equal("val1", vals["key1"])
+	assert.Equal(int64(2), vals["key2"])
+	assert.Equal(true, vals["key3"])
+
+	vals, err = Decode(sa, map[string]enumspb.IndexedValueType{
+		"key1": enumspb.INDEXED_VALUE_TYPE_STRING,
+		"key4": enumspb.INDEXED_VALUE_TYPE_INT,
+		"key3": enumspb.INDEXED_VALUE_TYPE_BOOL,
+	})
+	assert.Error(err)
+	assert.True(errors.Is(err, ErrInvalidName))
+	assert.Len(sa.IndexedFields, 3)
+	assert.Equal("val1", vals["key1"])
+	assert.Equal(int64(2), vals["key2"])
+	assert.Equal(true, vals["key3"])
+
+	delete(sa.IndexedFields["key1"].Metadata, "type")
+	delete(sa.IndexedFields["key2"].Metadata, "type")
+	delete(sa.IndexedFields["key3"].Metadata, "type")
+
+	vals, err = Decode(sa, nil)
+	assert.Error(err)
+	assert.True(errors.Is(err, ErrInvalidType))
+	assert.Len(vals, 3)
+	assert.Nil(vals["key1"])
+	assert.Nil(vals["key2"])
+	assert.Nil(vals["key3"])
 }
